Tidy up on demand tv tests
