+++ cypress/integration/pages/onDemandTV/testsForCanonicalOnly.js
@@ -1,40 +1,7 @@
-import appConfig from '../../../../src/server/utilities/serviceConfigs';
 import envConfig from '../../../support/config/envs';
-import {
-  getEmbedUrl,
-  isExpired,
-  dataEndpointOverride,
-} from '../../../support/helpers/onDemandRadioTv';
 
-export default ({ service, pageType, variant }) => {
+export default ({ service, pageType }) => {
   describe(`testsForCanonicalOnly for ${service} ${pageType}`, () => {
-    describe('AV Player', () => {
-      const { lang } = appConfig[service][variant];
-      let embedUrl;
-      let isExpiredEpisode;
-
-      it('should render an iframe with a valid URL', () => {
-        cy.request(
-          `${Cypress.env('currentPath')}.json${dataEndpointOverride()}`,
-        ).then(({ body: jsonData }) => {
-          embedUrl = getEmbedUrl(jsonData, lang);
-          isExpiredEpisode = isExpired(jsonData);
-
-          if (!isExpiredEpisode) {
-            // Ensure that the iFrame src matches the expected embed URL
-            cy.get('iframe').then(iframe => {
-              cy.wrap(iframe.prop('src')).should('contain', embedUrl);
-            });
-
-            cy.get(`iframe[src*="${embedUrl}"]`).should('be.visible');
-            cy.testResponseCodeAndType(embedUrl, 200, 'text/html');
-          } else {
-            cy.log(`Episode is expired: ${Cypress.env('currentPath')}`);
-          }
-        });
-      });
-    });
-
     describe('Chartbeat', () => {
       if (envConfig.chartbeatEnabled) {
         it('should have a script with src value set to chartbeat source', () => {
